Change location inside esphome folder
Recent changes to ESPHome makes it impossible to use symbolic
links for the yaml files. This causes the builds to fail
silently. As such we need to move this subproject inside
the ESPHome folder. Given that the ESPhome only loads yaml
files from the root of esphome, we need to create a yaml file
there loading the real one. ESPHome also requires a secrets.yaml
but does not find it when stored at the original place, hence
we create a symbolic link to where the file is expected to be.