Use UART definition from yaml file
The C++ code used the Serial2 API for communicating
with the heating boiler, while there was also a
definition in the YAML file. That definition is not
really used and overwritten by the Serial2 usage
in the C++ code.
We now use the actual component in the YAML file
for controlling the UART and stop using the lower
level API for Serial2.